pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ankole-framework"
version = "2.0.0"
description = "Multi-driver E2E test framework: Playwright, pywinauto, httpx, subprocess — with full observability stack"
requires-python = ">=3.10"
dependencies = [
    "pytest>=7.4.0,<9",
    "pytest-ordering>=0.6",
    "pytest-timeout>=2.2.0,<3",
    "allure-pytest>=2.13.0,<3",
    "Pillow>=10.0.0,<12",
    "mss>=9.0.0,<10",
    "PyYAML>=6.0.1,<7",
    "python-dotenv>=1.0.0,<2",
    "tcms-api>=12.0,<14",
]

[project.optional-dependencies]
# Windows-only: UI automation
desktop = ["pywinauto>=0.6.8"]
# Web UI automation (Playwright)
web = ["playwright>=1.40"]
# API testing
api = ["httpx>=0.25"]
# Remote agent / Loki log collection
remote = ["requests>=2.28,<3"]
# Grafana/Prometheus metrics
metrics = ["prometheus-client>=0.20.0,<1"]
# Security testing (OWASP ZAP + JWT)
security = ["python-owasp-zap-v2.4>=0.0.21", "PyJWT>=2.8.0"]
# Load testing (Locust)
load = ["locust>=2.20", "gevent>=23.0"]
# Test data factories (Faker-powered)
factory = ["factory-boy>=3.3.0", "Faker>=22.0"]
# Database testing (PostgreSQL)
db = ["psycopg2-binary>=2.9"]
# API schema validation
schema = ["pydantic>=2.0", "jsonschema>=4.20"]
# Visual regression testing
visual = ["pixelmatch>=0.3.0"]
# API mocking
mock = ["respx>=0.21"]
# Parallel execution
parallel = ["pytest-xdist>=3.5"]
# All optional dependencies
all = [
    "pywinauto>=0.6.8",
    "playwright>=1.40",
    "httpx>=0.25",
    "prometheus-client>=0.20.0,<1",
    "requests>=2.28,<3",
    "python-owasp-zap-v2.4>=0.0.21",
    "PyJWT>=2.8.0",
    "locust>=2.20",
    "gevent>=23.0",
    "factory-boy>=3.3.0",
    "Faker>=22.0",
    "psycopg2-binary>=2.9",
    "pydantic>=2.0",
    "jsonschema>=4.20",
    "pixelmatch>=0.3.0",
    "respx>=0.21",
    "pytest-xdist>=3.5",
]

# Auto-register pytest plugin so consumer repos get fixtures/hooks for free
[project.entry-points.pytest11]
ankole = "ankole.plugin"

[tool.setuptools.packages.find]
include = ["ankole*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "apps: Application that used in testing purpose",
    "web: Web UI automation tests (Playwright)",
    "api: REST API tests (httpx)",
    "cli: CLI tool tests (subprocess)",
    "desktop: Windows desktop UI automation tests (pywinauto)",
    "smoke: Quick smoke tests",
    "regression: Full regression tests",
    "critical: Critical path tests that must pass",
    "slow: Tests that take longer than 30 seconds",
    "tcms(case_id): Map this test to a specific Kiwi TCMS TestCase ID",
    "order(index): Execution order for pytest-ordering plugin",
    "depends_on(*case_ids): Skip test if any dependency TC did not pass",
    "flow: Flow-based tests (use -m flow to run only flow tests, -m 'not flow' for imperative only)",
    "security: OWASP security tests",
    "owasp: OWASP Top 10 vulnerability tests",
    "load: Load and performance tests",
    "zap: Tests requiring OWASP ZAP proxy",
    "a11y: Accessibility tests (axe-core)",
    "visual: Visual regression tests",
    "quarantine: Flaky tests quarantined from CI (skipped by default)",
]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--color=no",
    "--alluredir=evidence/allure-results",
]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)s] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
timeout = 600
timeout_method = "thread"
filterwarnings = [
    "ignore:Revert to STA COM threading mode:UserWarning",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
